fix: add deferred contruction to hardware interfaces
